code fonctionnel apres restructuration du projet

+++ src/vecteur.cpp
@@ -1,3 +1,3 @@
 #include "vecteur.hpp"
 
-Vecteur::Vecteur(Point position, int w, int h) : position(position), w(w), h(h) {}
+Vecteur::Vecteur(Point position, float w, float h) : position(position), w(w), h(h) {}
